reorganize the training code
